Process a single .agd file

The function analyses a .agd file using the functions from the activAnalyzer R package and using information relating to the demographic characteristics of the participant, the dates (start and end) of the physical activity measurement (optional), and the processing settings.

process_agd(
  agd_file,
  config,
  demo,
  dates = NULL,
  id_config,
  content = c("option_1", "option_2", "option_3")
)

Arguments

agd_file

A character string to set the path to the .agd file. The name of the file shoud begin with the participant number id (e.g., "001").

config

A dataframe obtained after importing the config.csv file needed to allow the analysis process. Use the get_config_file function to get the default config.csv file.

demo

A dataframe containing the demographic variables needed to compute physical activity metrics, that are:

id: participant inclusion number.

age: age of the participant.

sex: sex of the participant. Males could be indicated by male, and females by female.

weight: weight of the participant in kg.

The names of these variables can be differently written in your own file containing these variables and should be set in the config.csv file to be filled to allow the analysis process. Use the get_demo_template function to get a template of the file to be imported with the required demographic characteristics.

dates

A dataframe containing the start and end dates of the physical activity measurement. Use the get_dates_template function to get a template of the file to be imported with the required dates if wanted. Default is NULL. If NULL, all the days of measurement will be analysed. If an appropriate dataframe is provided, the analyses will be conducted using the dates provided in the dataframe.

id_config

A numeric value to indicate how many numbers are used to identify the participant in the name of the .agd file (e.g., 2 for "01", 3 for "001", 4 for "0001", etc). The name of the file to be analysed has to begin by the participant number written as in the mentionned examples above.

content

A character value to indicate what results should be contained in the list of the exported results:

option_1 provides only the metrics summarised using valid days.

option_2 provides, in addition to option_1, the results computed for each day of measurement.

option_3 provides, in addition to option_2, the initial dataset with all the marks of the computations performed during the analysis process.

Value

The datasets required by the content argument of the function:

all_metrics: the metrics summarized using the valid days.

results_by_day: the results computed for each day of measurement (option_2 and option_3 only).

data_with_intensity_marks: the initial dataset with all the marks of the computations performed during the analysis process (option_3 only).

If the content argument is option_1, a dataframe is returned. Otherwise a list is returned.

Examples

FILE_PATH <- system.file("agd/001.agd", package = "activAnalyzer.batch")
CONFIG <- readr::read_csv2(system.file("config.csv", package = "activAnalyzer.batch"))
#> ℹ Using "','" as decimal and "'.'" as grouping mark. Use `read_delim()` for more control.
#> Rows: 55 Columns: 2
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ";"
#> chr (2): CODE_NAME, USER_VALUE
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
DEMO <- readr::read_csv2(system.file("DEMO.csv", package = "activAnalyzer.batch"))
#> ℹ Using "','" as decimal and "'.'" as grouping mark. Use `read_delim()` for more control.
#> Rows: 16 Columns: 4
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ";"
#> chr (1): sex
#> dbl (3): id, age, weight
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
DATES <- readr::read_csv2(system.file("DATES.csv", package = "activAnalyzer.batch"))
#> ℹ Using "','" as decimal and "'.'" as grouping mark. Use `read_delim()` for more control.
#> Rows: 10 Columns: 3
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ";"
#> chr (2): start_date, end_date
#> dbl (1): id
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
process_agd(
agd_file = FILE_PATH, 
config = CONFIG, 
demo = DEMO, 
dates = DATES, id_config = 3
)
#> frame is 60
#> streamFrame is 0
#> allowanceFrame is 0
#> You have computed intensity metrics with the mark_intensity() function using the following inputs: 
#>     axis = vm
#>     sed_cutpoint = 150 counts/min
#>     mpa_cutpoint = 3208 counts/min
#>     vpa_cutpoint = 8565 counts/min
#>     equation = Santos-Lozano et al. (2013) [Adults]
#>     age = 56
#>     weight = 98
#>     sex = male
#> Joining with `by = join_by(date)`
#> Joining with `by = join_by(date)`
#> Joining with `by = join_by(date)`
#> You have computed results with the recap_by_day() function using the following inputs: 
#>          age = 56
#>          weight = 98
#>          sex = male
#>   id valid_days wear_time total_counts_axis1 total_counts_vm axis1_per_min
#> 1  1          5     767.8           513108.6        970344.6        686.68
#>   vm_per_min minutes_SED minutes_LPA minutes_MPA minutes_VPA minutes_MVPA
#> 1    1290.55       260.8       435.4        53.2        18.4         71.6
#>   percent_SED percent_LPA percent_MPA percent_VPA percent_MVPA ratio_mvpa_sed
#> 1       34.06       56.28         7.1        2.56         9.67           0.27
#>   mets_hours_mvpa total_kcal  pal total_steps max_steps_60min max_steps_30min
#> 1            3.03    1945.72 1.09       14869           57.85           70.05
#>   max_steps_20min max_steps_5min max_steps_1min peak_steps_60min
#> 1           74.29           95.2            109             74.6
#>   peak_steps_30min peak_steps_20min peak_steps_5min peak_steps_1min    ig
#> 1             86.6            91.75          105.24             109 -1.51
#>      M1/3   M120     M60     M30     M15      M5 mean_breaks alpha MBD  UBD
#> 1 1983.47 4330.7 5123.32 5830.03 6311.44 6824.47        68.8  2.28   2 6.78
#>   gini
#> 1 0.57