Process all .agd files

The function analyses all the .agd files present in a given directory using the functions from the activAnalyzer R package and using information relating to the demographic characteristics of the participant(s), the dates (start and end) of the physical activity measurement(s) (optional), and the processing settings.

process_all_agd(
  agd_dir,
  config_path,
  demo_path,
  dates_path = NULL,
  id_config,
  content = c("option_1", "option_2", "option_3"),
  out_dir = "./out",
  export_results = FALSE,
  cores = NULL
)

Arguments

agd_dir

A character string to set the path to the directory where are all the .agd files to be analysed. The names of the files shoud begin with the participant number id (e.g., "001").

config_path

A character string to set the path to the config.csv file to be filled to allow the analysis process. Use the get_config_file function to get the default config.csv file.

demo_path

A character string to set the path to the .csv file that should contain the demographic variables needed to compute physical activity metrics, that are:

id: participant inclusion number.

age: age of the participant.

sex: sex of the participant. Males could be indicated by male, and females by female.

weight: weight of the participant in kg.

The names of these variables can be differently written in your own file containing these variables and should be set in the config.csv file to be filled to allow the analysis process. Use the get_demo_template function to get a template of the file to be imported with the required demographic characteristics.

dates_path

A character string to set the path to the .csv file that contains the start and end dates of the physical activity measurement(s). Use the get_dates_template function to get a template of the file to be imported with the required dates if wanted. Default is NULL. If NULL, all the days of measurement will be analysed. If the path to an appropriate file is provided, the analyses will be conducted using the dates provided in the file.

id_config

A numeric value to indicate how many numbers are used to identify the participant in the names of the .agd files (e.g., 2 for "01", 3 for "001", 4 for "0001", etc). The names of the files to be analysed have to begin by the participant number written as in the mentionned examples above.

content

A character value to indicate what results should be contained in the list of the exported results:

option_1 provides only the metrics summarised using valid days.

option_2 provides, in addition to option_1, the results computed for each day of measurement.

option_3 provides, in addition to option_2, the initial dataset(s) with all the marks of the computations performed during the analysis process.

out_dir

A character string to set the path to the directory that will be used to export the results if required.

export_results

A logical value regarding the export of the results. The results are exported as a dataframe (if content argument = "option_1") or as a list object otherwise, embedded in a .rds file. The results exported are the followings:

option_1: the metrics summarized using the valid days only.

option_2: the results from option_1 plus the results computed for each day of measurement.

option_3: the results from option_2 plus the initial dataset(s) with all the marks of the computations performed during the analysis process.

cores

A positive integer representing the number of cores to use for parallel processing (i.e., the number of processes to start). This value must be between 2 and parallel::detectCores() - 1. Default is parallel::detectCores() - 1.

Value

The data required by the content argument of the function:

all_metrics: the metrics summarized using the valid days and collapsed into a single dataframe.

results_by_day: the results computed for each day of measurement collapsed into a single dataframe (option_2 and option_3 only).

data_with_intensity_marks: the initial dataset(s) with all the marks of the computations performed during the analysis process collapsed into a single dataframe (option_3 only).

If the content argument is option_1, a dataframe is returned. Otherwise a list of dataframes is returned.

Examples

DIR_AGD_PATH <- system.file("agd/", package = "activAnalyzer.batch")
DEMO_PATH <- system.file("DEMO.csv", package = "activAnalyzer.batch")
DATES_PATH <- system.file("DATES.csv", package = "activAnalyzer.batch")
CONFIG_PATH <- system.file("config.csv", package = "activAnalyzer.batch")
process_all_agd(
agd_dir = DIR_AGD_PATH, 
demo_path = DEMO_PATH, 
config_path = CONFIG_PATH,
dates_path = DATES_PATH, 
id_config = 3,
cores = 2
)
#> Error in process_all_agd(agd_dir = DIR_AGD_PATH, demo_path = DEMO_PATH,     config_path = CONFIG_PATH, dates_path = DATES_PATH, id_config = 3,     cores = 2): unused argument (cores = 2)